+++ operators/markers_as_timecodes.py
@@ -33,15 +33,20 @@ class POWER_SEQUENCER_OT_copy_markers_as_timecodes(bpy.types.Operator):
             self.report({"INFO"}, "No markers found")
             return {"CANCELLED"}
 
-        sorted_markers = sorted(context.scene.timeline_markers, key=op.attrgetter("frame"))
+        sorted_markers = sorted(context.scene.timeline_markers,
+                                key=lambda m: m.frame)
+
+        framerate = render.fps / render.fps_base
+        last_marker_seconds = sorted_markers[-1].frame / framerate
+        seconds_in_hour = 3600.0
+        time_format = "%H:%M:%S" if last_marker_seconds >= seconds_in_hour else "%M:%S"
 
         markers_as_timecodes = []
         for marker in sorted_markers:
-            framerate = render.fps / render.fps_base
             time = dt.datetime(year=1, month=1, day=1) + dt.timedelta(
-                seconds=marker.frame / framerate
-            )
-            markers_as_timecodes.append(time.strftime("%H:%M:%S {}".format(marker.name)))
+                seconds=marker.frame / framerate)
+            markers_as_timecodes.append(
+                time.strftime(time_format) + " " + marker.name)
 
         pyperclip.copy("\n".join(markers_as_timecodes))
         return {"FINISHED"}